Refactor

+++ Core/src/Serialization/Serialization.hpp
@@ -26,9 +26,9 @@ namespace Pixf::Core::Serialization {
     inline vec3 DeserializeVec3(Json::object json) {
         vec3 vector;
 
-        vector.x = json["x"].to_number<double>();
-        vector.y = json["y"].to_number<double>();
-        vector.z = json["z"].to_number<double>();
+        vector.x = json["x"].to_number<float>();
+        vector.y = json["y"].to_number<float>();
+        vector.z = json["z"].to_number<float>();
 
         return vector;
     }
@@ -44,10 +44,10 @@ namespace Pixf::Core::Serialization {
     inline vec4 DeserializeVec4(Json::object json) {
         vec4 vector;
 
-        vector.x = json["x"].to_number<double>();
-        vector.y = json["y"].to_number<double>();
-        vector.z = json["z"].to_number<double>();
-        vector.w = json["w"].to_number<double>();
+        vector.x = json["x"].to_number<float>();
+        vector.y = json["y"].to_number<float>();
+        vector.z = json["z"].to_number<float>();
+        vector.w = json["w"].to_number<float>();
 
         return vector;
     }
@@ -63,10 +63,10 @@ namespace Pixf::Core::Serialization {
     inline quat DeserializeQuat(Json::object json) {
         quat quat;
 
-        quat.x = json["x"].to_number<double>();
-        quat.y = json["y"].to_number<double>();
-        quat.z = json["z"].to_number<double>();
-        quat.w = json["w"].to_number<double>();
+        quat.x = json["x"].to_number<float>();
+        quat.y = json["y"].to_number<float>();
+        quat.z = json["z"].to_number<float>();
+        quat.w = json["w"].to_number<float>();
 
         quat = normalize(quat);
 
@@ -83,9 +83,9 @@ namespace Pixf::Core::Serialization {
     inline vec3 DeserializeColorRgb(Json::object json) {
         vec3 vector;
 
-        vector.r = json["r"].to_number<double>();
-        vector.g = json["g"].to_number<double>();
-        vector.b = json["b"].to_number<double>();
+        vector.r = json["r"].to_number<float>();
+        vector.g = json["g"].to_number<float>();
+        vector.b = json["b"].to_number<float>();
 
         return vector;
     }
@@ -101,10 +101,10 @@ namespace Pixf::Core::Serialization {
     inline vec4 DeserializeColorRgba(Json::object json) {
         vec4 vector;
 
-        vector.r = json["r"].to_number<double>();
-        vector.g = json["g"].to_number<double>();
-        vector.b = json["b"].to_number<double>();
-        vector.a = json["a"].to_number<double>();
+        vector.r = json["r"].to_number<float>();
+        vector.g = json["g"].to_number<float>();
+        vector.b = json["b"].to_number<float>();
+        vector.a = json["a"].to_number<float>();
 
         return vector;
     }
